Simplification efforts
-Removed unneccesary null checks all over the place with lateinit
-Naive automatic detection of if a topic is a pattern
-Attempted to simplify controller

I removed the need for the controller to know about the camel context,
but was unable to remove it completely so far. The controller
shouldn't need to know anything about camel or the routes,
with the exception of the dynamic routes that are created
when adding topics.

I keep thinking I should be using injection, but that is just more
cognitive load.